# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OSS-register API v1",
    "version": "1.0.0",
    "contact": {
      "email": "developer.overheid@geonovum.nl",
      "name": "Team developer.overheid.nl",
      "url": "https://github.com/developer-overheid-nl/don-oss-register/issues"
    },
    "description": "API to access the OSS register of developer.overheid.nl.\n\n## Auth\n\nThis API distinguishes between public and private endpoints.\nPublic endpoints can be accessed with either an API key or a client credentials token.\nPrivate endpoints can only be accessed with a client credentials token.\n\n### API key\n\nUsing an API key, you can access all public endpoints of the API register.\nThese requests can also be made from the browser.\nSimply pass the obtained API key with each request using the `X-Api-Key` header.\n\n### Client credentials token\n\nUsing a client credentials token, you can access both public and private endpoints of the OSS register.\nTo obtain the token, you need to perform a `POST` request to `https://auth.developer.overheid.nl/realms/don/protocol/openid-connect/token` with the following Form URL Encoded body:\n- `grant_type`: `client_credentials`\n- `scope`: depending on the access you need and the client you are, you can request one or more of the following scopes:\n  - `repositories:read`\n  - `repositories:write`\n  - `gitOrganisations:read`\n  - `gitOrganisations:write`\n  - `organisations:read`\n  - `organisations:write`\n- `client_id`: the client id you received from us\n- `client_secret`: the client secret you received from us\n\nPass the obtained token with each request using the `Authorization` header. Example:\n\n`Authorization`: `Bearer {ACCESS_TOKEN}` (replace `{ACCESS_TOKEN}` with the obtained `access_token`)\n\n## Pagination\n\nPagination of collections is done using the `Link` header.\nThere are various libraries available (such as [parse-link-header](https://www.npmjs.com/package/parse-link-header) for Javascript) that can parse this header.\nAdditionally, the following headers provide extra support for implementing pagination in the client:\n- `Current-Page`: the current page in the collection\n- `Per-Page`: the number of items per page\n- `Total-Count`: the total number of items\n- `Total-Pages`: the total number of pages\n"
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.developer.overheid.nl/oss-register/v1"
    },
    {
      "description": "Test",
      "url": "https://api-register.don.apps.digilab.network/oss-register/v1"
    }
  ],
  "tags": [
    {
      "name": "Repositories",
      "description": "Endpoints for listing and managing repositories."
    },
    {
      "name": "Organisations",
      "description": "Endpoints for listing and managing organisations."
    },
    {
      "name": "Git organisations",
      "description": "Endpoints for listing and managing git organisations."
    },
    {
      "name": "Public endpoints",
      "description": "Public endpoints, accessible with an API key or client credentials token."
    },
    {
      "name": "Private endpoints",
      "description": "Private endpoints of the API register, accessible with a client credentials token."
    }
  ],
  "paths": {
    "/git-organisations": {
      "get": {
        "tags": [
          "Public endpoints",
          "Git organisations"
        ],
        "summary": "List git organisations",
        "description": "Geeft een lijst terug met git organisations die in het register zijn opgenomen.",
        "operationId": "listGitOrganisations",
        "parameters": [
          {
            "$ref": "./components/parameters.yaml#/page"
          },
          {
            "$ref": "./components/parameters.yaml#/perPage"
          },
          {
            "$ref": "./components/parameters.yaml#/organisation"
          }
        ],
        "responses": {
          "400": {
            "$ref": "./components/responses.yaml#/400"
          },
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              },
              "Link": {
                "$ref": "./components/headers.yaml#/Link"
              },
              "Total-Count": {
                "$ref": "./components/headers.yaml#/Total-Count"
              },
              "Current-Page": {
                "$ref": "./components/headers.yaml#/Current-Page"
              },
              "Per-Page": {
                "$ref": "./components/headers.yaml#/Per-Page"
              },
              "Total-Pages": {
                "$ref": "./components/headers.yaml#/Total-Pages"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/GitOrganisation.schema.json"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Private endpoints",
          "Git organisations"
        ],
        "summary": "Create git organisation",
        "description": "Registreer een nieuwe git organisatie in het register.",
        "operationId": "createGitOrganisation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/GitOrganisationInput.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/GitOrganisation.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "./components/responses.yaml#/400"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "gitOrganisations:write"
            ]
          }
        ]
      }
    },
    "/repositories": {
      "get": {
        "tags": [
          "Public endpoints",
          "Repositories"
        ],
        "summary": "List repositories",
        "description": "Geeft een lijst terug met OSS repositories die in het register zijn opgenomen.",
        "operationId": "listRepositories",
        "parameters": [
          {
            "$ref": "./components/parameters.yaml#/page"
          },
          {
            "$ref": "./components/parameters.yaml#/perPage"
          },
          {
            "$ref": "./components/parameters.yaml#/organisation"
          }
        ],
        "responses": {
          "400": {
            "$ref": "./components/responses.yaml#/400"
          },
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              },
              "Link": {
                "$ref": "./components/headers.yaml#/Link"
              },
              "Total-Count": {
                "$ref": "./components/headers.yaml#/Total-Count"
              },
              "Current-Page": {
                "$ref": "./components/headers.yaml#/Current-Page"
              },
              "Per-Page": {
                "$ref": "./components/headers.yaml#/Per-Page"
              },
              "Total-Pages": {
                "$ref": "./components/headers.yaml#/Total-Pages"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/RepositorySummary.schema.json"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Private endpoints",
          "Repositories"
        ],
        "summary": "Create repository",
        "description": "Registreer een nieuwe OSS repository in het register.",
        "operationId": "createRepository",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/RepositoryInput.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Repository.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "./components/responses.yaml#/400"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "repositories:write"
            ]
          }
        ]
      }
    },
    "/repositories/_search": {
      "get": {
        "tags": [
          "Public endpoints",
          "Repositories"
        ],
        "summary": "Search repositories",
        "description": "Geeft een lijst terug met OSS repositories die in het register zijn opgenomen.",
        "operationId": "searchRepositories",
        "parameters": [
          {
            "$ref": "./components/parameters.yaml#/page"
          },
          {
            "$ref": "./components/parameters.yaml#/perPage"
          },
          {
            "$ref": "./components/parameters.yaml#/organisation"
          },
          {
            "$ref": "./components/parameters.yaml#/q"
          }
        ],
        "responses": {
          "400": {
            "$ref": "./components/responses.yaml#/400"
          },
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              },
              "Link": {
                "$ref": "./components/headers.yaml#/Link"
              },
              "Total-Count": {
                "$ref": "./components/headers.yaml#/Total-Count"
              },
              "Current-Page": {
                "$ref": "./components/headers.yaml#/Current-Page"
              },
              "Per-Page": {
                "$ref": "./components/headers.yaml#/Per-Page"
              },
              "Total-Pages": {
                "$ref": "./components/headers.yaml#/Total-Pages"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/RepositorySummary.schema.json"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/repositories/{id}": {
      "parameters": [
        {
          "$ref": "./components/parameters.yaml#/id"
        }
      ],
      "get": {
        "security": [
          {
            "apiKey": []
          },
          {
            "clientCredentials": [
              "repositories:read"
            ]
          }
        ],
        "tags": [
          "Public endpoints",
          "Repositories"
        ],
        "summary": "Get repository by id",
        "description": "Geeft \u00e9\u00e9n OSS repository terug op basis van het id.",
        "operationId": "getRepositoryById",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Repository.schema.json"
                }
              }
            }
          },
          "404": {
            "$ref": "./components/responses.yaml#/404"
          }
        }
      },
      "put": {
        "security": [
          {
            "clientCredentials": [
              "repositories:write"
            ]
          }
        ],
        "tags": [
          "Private endpoints",
          "Repositories"
        ],
        "summary": "Specifieke repository updaten",
        "description": "Specifieke repository updaten",
        "operationId": "updateRepository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/RepositoryInput.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Repository.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "./components/responses.yaml#/400"
          },
          "404": {
            "$ref": "./components/responses.yaml#/404"
          }
        }
      }
    },
    "/organisations": {
      "get": {
        "security": [
          {
            "apiKey": []
          },
          {
            "clientCredentials": [
              "organisations:read"
            ]
          }
        ],
        "tags": [
          "Public endpoints",
          "Organisations"
        ],
        "summary": "Alle organisaties ophalen",
        "description": "Alle organisaties ophalen",
        "operationId": "listOrganisations",
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/OrganisationSummary.schema.json"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "clientCredentials": [
              "organisations:write"
            ]
          }
        ],
        "tags": [
          "Private endpoints",
          "Organisations"
        ],
        "summary": "Organisatie aanmaken",
        "description": "Maak een nieuwe organisatie aan.",
        "operationId": "createOrganisation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/OrganisationSummary.schema.json"
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "./components/responses.yaml#/400"
          },
          "201": {
            "headers": {
              "API-Version": {
                "$ref": "./components/headers.yaml#/API-Version"
              }
            },
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Organisation.schema.json"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "$ref": "./components/securitySchemes.yaml#/apiKey"
      },
      "clientCredentials": {
        "$ref": "./components/securitySchemes.yaml#/clientCredentials"
      }
    }
  },
  "security": [
    {
      "clientCredentials": []
    }
  ]
}